use tcp-nodelay for WebSocket
Setting TCP_NODELAY is useful on Linux systems where some legitimate
transmission patterns run afoul of Nagle's algorithm. On macOS 10.5 and
later, which implement Minshall's modification to Nagle's algorithm,
setting this parameter is not necessary, but does not appear to hurt.

For a nice explanation, see http://www.stuartcheshire.org/papers/NagleDelayedAck.

Chez Scheme's max-put-copy provokes Nagle's algorithm for certain
WebSocket payloads where we write the message header in small pieces
that go to the port's buffer followed by a payload that causes the
custom port to flush the buffered header and write the payload with
separate calls to the OS.

Yet even when we write the header and payload in a single OS call
there are cases where multiple calls to ws:send without waiting for
a ws:message ran afoul of Nagle's algorithm. Setting TCP_NODELAY
resolved those cases.