WIP ci: add support for Chez Scheme 10.x (move this)
The first job in the workflow ("build") ensures that we have a cache of
the .git directory and boot files for a6le, a6osx, and a6nt built from
the "main" Chez Scheme branch. This job uses the v9.6.4 a6le boot file
to bootstrap via the re.boot target to avoid the overhead of building
via pb on multiple platforms.

The second job in the workflow ("mats") tests a matrix of system
configurations. This job checks out the Swish repository and then
restores the cached Chez Scheme boot files. For the "main" Chez Scheme
branch, the job uses the --force flag to configure and builds the C
kernel. For tagged releases v9.6.4 and earlier, the boot files are
already checked in, so again we configure and build only the C side.

In either case, we set the PATH and SCHEMEHEAPDIRS so that Swish's
configure script finds the "scheme" executable in PATH and the build and
test process continues as before.